src/requirements.txt
google-genai==1.59.0
fastapi==0.128.0
uvicorn==0.40.0
PyPDF2==3.0.1
python-multipart==0.0.21
pydantic-settings==2.12.0
requests==2.32.5

